Restarted Fire
Added all the scripts and keyid, and I couldn't get the fire to work right. It started deactivated and I couldn't get it to do anything else. Deleted the fire sprite completely to start over. Added a new one with the FireActive animation and FireDeactivate animation. We'll get em' next itme