Attached On Trigger
Alright we're getting somewhere. Added the fire prefab, and tried to mess with the match events (to no avail).
However I removed the triggerbehaviour script from the melon, and gave it the attach on trigger which is now letting the player carry it!! (It still doesn't affect the fire yet.)
Also put all the new scripts back to their vanilla provided scripts.
Made a new key and deleted the old one