Prefabs don't work
okay we have a mess. I deleted the pink guy because I couldn't get him to navigate in the same area as the masked character? (I would like to have him jump off the ledge when he sees the character so people know they should jump down.)

Also now the spikes are doing the same thing the strawberries were doing (Not triggering despite being a commit.) So we're gonna try and make their elements prefabs, so it works.